refactor(sink): Phase 2 — unified SinkFormat trait + absorb FormatSeeder
Introduce SinkFormat in io/write/sink_format.rs: a single trait combining
write() (formerly AcceptedSinkAdapter) and seed_unique_tracker() (formerly
FormatSeeder). A static SINK_FORMATS registry replaces the match arms in
both dispatch sites.

Format files (delta, parquet, iceberg) each implement SinkFormat once;
seeder logic moves in from the deleted unique_seed/{delta,parquet,iceberg}.rs.
unique_seed/mod.rs becomes a thin dispatcher via sink_format(name)?.seed_unique_tracker().
AcceptedSinkAdapter trait and accepted_sink_adapter() fn are deleted.
Runtime::accepted_sink_adapter() → Runtime::sink_format().
